refactor: unify prototype setups and split tool hooks
