+++ packages/prototypes/base/src/button/index.ts
@@ -1,14 +1,8 @@
-import { definePrototype } from '@proto.ui/core';
-import { asButton } from './as-button';
-
-export type { ButtonProps, ButtonExposes, ButtonStateHandles, ButtonAsHookContract } from './types';
-export { asButton } from './as-button';
-
-const button = definePrototype({
-  name: 'base-button',
-  setup: () => {
-    asButton();
-  },
-});
-
-export default button;
+export { asButton } from './button';
+export { default } from './button';
+export type {
+  ButtonProps,
+  ButtonExposes,
+  ButtonStateHandles,
+  ButtonAsHookContract,
+} from './button';
