+++ packages/prototypes/base/src/dropdown/index.ts
@@ -1,7 +1,4 @@
-import dropdownContent from './content';
-import dropdownItem from './item';
 import dropdownRoot from './root';
-import dropdownTrigger from './trigger';
 
 export type {
   DropdownContentAsHookContract,
@@ -20,10 +17,9 @@ export type {
   DropdownTriggerProps,
 } from './types';
 
-export { asDropdownRoot } from './as-dropdown-root';
-export { asDropdownTrigger } from './as-dropdown-trigger';
-export { asDropdownContent } from './as-dropdown-content';
-export { asDropdownItem } from './as-dropdown-item';
-export { dropdownRoot, dropdownTrigger, dropdownContent, dropdownItem };
+export { asDropdownRoot, default as dropdownRoot } from './root';
+export { asDropdownTrigger, default as dropdownTrigger } from './trigger';
+export { asDropdownContent, default as dropdownContent } from './content';
+export { asDropdownItem, default as dropdownItem } from './item';
 
 export default dropdownRoot;
