+++ packages/prototypes/base/src/hover-card/index.ts
@@ -1,6 +1,4 @@
-import hoverCardContent from './content';
 import hoverCardRoot from './root';
-import hoverCardTrigger from './trigger';
 
 export type {
   HoverCardContentAsHookContract,
@@ -18,9 +16,8 @@ export type {
 export type { HoverCardContextValue } from './shared';
 
 export { HOVER_CARD_CONTEXT, HOVER_CARD_FAMILY } from './shared';
-export { asHoverCardRoot } from './as-hover-card-root';
-export { asHoverCardTrigger } from './as-hover-card-trigger';
-export { asHoverCardContent } from './as-hover-card-content';
-export { hoverCardRoot, hoverCardTrigger, hoverCardContent };
+export { asHoverCardRoot, default as hoverCardRoot } from './root';
+export { asHoverCardTrigger, default as hoverCardTrigger } from './trigger';
+export { asHoverCardContent, default as hoverCardContent } from './content';
 
 export default hoverCardRoot;
