+++ packages/prototypes/base/src/tabs/index.ts
@@ -1,8 +1,3 @@
-import tabsContent from './content';
-import tabsList from './list';
-import tabsRoot from './root';
-import tabsTrigger from './trigger';
-
 export type {
   TabsRootProps,
   TabsRootExposes,
@@ -23,8 +18,7 @@ export type {
 export type { TabsActivationMode, TabsContextValue, TabsOrientation } from './shared';
 
 export { TABS_CONTEXT, TABS_FAMILY } from './shared';
-export { asTabsRoot } from './as-tabs-root';
-export { asTabsList } from './as-tabs-list';
-export { asTabsTrigger } from './as-tabs-trigger';
-export { asTabsContent } from './as-tabs-content';
-export { tabsRoot, tabsList, tabsTrigger, tabsContent };
+export { asTabsRoot, default as tabsRoot } from './root';
+export { asTabsList, default as tabsList } from './list';
+export { asTabsTrigger, default as tabsTrigger } from './trigger';
+export { asTabsContent, default as tabsContent } from './content';
