+++ packages/prototypes/base/src/tools/index.ts
@@ -1,4 +1,4 @@
-export { asEscapeKey } from './as-escape-key';
-export type { EscapeKeyOptions } from './as-escape-key';
-export { asOpenState } from './as-open-state';
-export type { OpenStateExposes, OpenStateHandles, OpenStateOptions } from './as-open-state';
+export { useEscapeKey } from './use-escape-key';
+export type { EscapeKeyOptions } from './use-escape-key';
+export { useOpenState } from './use-open-state';
+export type { OpenStateExposes, OpenStateHandles, OpenStateOptions } from './use-open-state';
