+++ packages/prototypes/shadcn/README.md
@@ -28,9 +28,6 @@ npm install @proto.ui/prototypes-shadcn@0.0.1
 ## Related Internal Packages
 
 - `@proto.ui/core`
-- `@proto.ui/module-as-trigger`
-- `@proto.ui/module-event`
-- `@proto.ui/module-rule-meta`
 - `@proto.ui/prototypes-base`
 
 ## License
